Bench: 14469242

+++ src/util/pretty.hpp
@@ -8,18 +8,18 @@
 
 namespace Clockwork {
 
-inline void print_progress(size_t current, size_t total, f32 loss, size_t bar_width = 40) {
+inline void print_progress(size_t current, size_t total, f64 loss, size_t bar_width = 40) {
     if (total == 0) {
         std::cout << "\r[error: total=0]" << std::flush;
         return;
     }
 
-    f32 progress = static_cast<f32>(current) / static_cast<f32>(total);
-    if (progress > 1.0f) {
-        progress = 1.0f;
+    f64 progress = static_cast<f64>(current) / static_cast<f64>(total);
+    if (progress > 1.0) {
+        progress = 1.0;
     }
 
-    size_t pos = static_cast<size_t>(static_cast<f32>(bar_width) * progress);
+    size_t pos = static_cast<size_t>(static_cast<f64>(bar_width) * progress);
 
     std::ostringstream loss_ss;
     loss_ss << std::fixed << std::setprecision(6) << loss;
@@ -35,7 +35,7 @@ inline void print_progress(size_t current, size_t total, f32 loss, size_t bar_wi
         }
     }
 
-    std::cout << "] " << static_cast<int>(progress * 100.0f) << "% (" << current << "/" << total
+    std::cout << "] " << static_cast<int>(progress * 100.0) << "% (" << current << "/" << total
               << ") loss=" << loss_ss.str() << std::flush;
 }
